+++ recipes/base.rb
@@ -11,20 +11,26 @@ def cl_interface(interface, data)
   cumulus_interface interface do
     ipv4 data.ipv4 if data['ipv4']
     ipv6 data.ipv6 if data['ipv6']
+    alias_name data.alias if data['alias']
     speed data.speed if data['speed']
     mtu data.mtu if data['mtu']
     post_up data.post_up if data['post_up']
+    pre_down data.pre_down if data['pre_down']
     addr_method data.addr_method if data['addr_method']
-    # Options for enabling clagd ports
+    virtual_mac data.virtual_mac if data['virtual_mac']
+    virtual_ip data.virtual_ip if data['virtual_ip']
+    vids data.vids if data['vids']
+    pvid data.pvid if data['pvid']
     clagd_enable data.clagd_enable if data['clagd_enable']
     clagd_peer_ip data.clagd_peer_ip if data['clagd_peer_ip']
     clagd_priority data.clagd_priority if data['clagd_priority']
     clagd_sys_mac data.clagd_sys_mac if data['clagd_sys_mac']
+    clagd_args data.clagd_args if data['clagd_args']
     mstpctl_portnetwork data.mstpctl_portnetwork unless data['mstpctl_portnetwork'].nil?
     mstpctl_portadminedge data.mstpctl_portadminedge unless data['mstpctl_portadminedge'].nil?
     mstpctl_bpduguard data.mstpctl_bpduguard unless data['mstpctl_bpduguard'].nil?
-    notifies :run, "execute[reload_networking]", :delayed
-    notifies :run, "execute[reload_loopback]", :delayed if interface =~ /^lo/
+    notifies :run, 'execute[reload_networking]', :delayed
+    notifies :run, 'execute[reload_loopback]', :delayed if interface =~ /^lo/
   end
 end
 
@@ -35,7 +41,7 @@ end
 
 # Interface ranges
 node.cumulus.interface_range.each do |range_str, data|
-  # range str should be something like "swp[1-24].100" or "swp[2-5]"
+  # range str should be something like 'swp[1-24].100' or 'swp[2-5]'
   range = range_str.match(/\[(\d+)-(\d+)\]/)
   (range[1]..range[2]).each do |id|
     ifname = range_str.gsub(/\[\d+-\d+\]/, id)
@@ -49,13 +55,29 @@ node.cumulus.bond.each do |bond, data|
     ipv4 data.ipv4 if data['ipv4']
     ipv6 data.ipv6 if data['ipv6']
     slaves data.slaves if data['slaves']
+    addr_method data.addr_method if data['addr_method']
+    alias_name data.alias if data['alias']
+    mtu data.mtu if data['mtu']
+    miimon data.miimon if data['miimon']
+    min_links data.min_links if data['min_links']
+    mode data.mode if data['mode']
+    xmit_hash_policy data.xmit_hash_policy if data['xmit_hash_policy']
     clag_id data.clag_id if data['clag_id']
+    lacp_rate data.lacp_rate if data['lacp_rate']
     lacp_bypass_allow data.lacp_bypass_allow if data['lacp_bypass_allow']
+    lacp_bypass_period data.lacp_bypass_period if data['lacp_bypass_period']
+    lacp_bypass_priority data.lacp_bypass_priority if data['lacp_bypass_priority']
+    lacp_bypass_all_active data.lacp_bypass_all_active if data['lacp_bypass_all_active']
+    virtual_mac data.virtual_mac if data['virtual_mac']
     virtual_ip data.virtual_ip if data['virtual_ip']
+    vids data.vids if data['vids']
+    pvid data.pvid if data['pvid']
+    post_up data.post_up if data['post_up']
+    pre_down data.pre_down if data['pre_down']
     mstpctl_portnetwork data.mstpctl_portnetwork unless data['mstpctl_portnetwork'].nil?
     mstpctl_portadminedge data.mstpctl_portadminedge unless data['mstpctl_portadminedge'].nil?
     mstpctl_bpduguard data.mstpctl_bpduguard unless data['mstpctl_bpduguard'].nil?
-    notifies :run, "execute[reload_networking]", :delayed
+    notifies :run, 'execute[reload_networking]', :delayed
   end
 end
 
@@ -65,10 +87,19 @@ node.cumulus.bridge.each do |bridge, data|
     ports data.ports if data['ports']
     ipv4 data.ipv4 if data['ipv4']
     ipv6 data.ipv6 if data['ipv6']
+    mtu data.mtu if data['mtu']
+    addr_method data.addr_method if data['addr_method']
+    alias_name data.alias if data['alias']
+    mstpctl_treeprio data.mstpctl_treeprio if data['mstpctl_treeprio']
+    post_up data.post_up if data['post_up']
+    pre_down data.pre_down if data['pre_down']
     virtual_ip data.virtual_ip if data['virtual_ip']
     virtual_mac data.virtual_mac if data['virtual_mac']
     stp data.stp unless data['stp'].nil?
-    notifies :run, "execute[reload_networking]", :delayed
+    vids data.vids if data['vids']
+    pvid data.pvid if data['pvid']
+    vlan_aware data.vlan_aware if data['vlan_aware']
+    notifies :run, 'execute[reload_networking]', :delayed
   end
 end
 
@@ -78,5 +109,5 @@ cumulus_ports 'speeds' do
   speed_40g node.cumulus.ports['40g'] if node.cumulus.ports['40g']
   speed_40g_div_4 node.cumulus.ports['40g_div_4'] if node.cumulus.ports['40g_div_4']
   speed_4_by_10g node.cumulus.ports['4_by_10g'] if node.cumulus.ports['4_by_10g']
-  notifies :restart, "service[switchd]", :delayed if node.cumulus.restart_switchd
+  notifies :restart, 'service[switchd]', :delayed if node.cumulus.restart_switchd
 end